asurdev/pyproject.toml
[project]
name = "asurdev-sentinel"
version = "3.2.0"
description = "Multi-agent financial decision system combining technical analysis with astrological timing"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-ollama>=0.1.0",
    "pyswisseph>=2.10.0",
    "astropy>=6.0.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "yfinance>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "sse-starlette>=2.0.0",
    "chromadb>=1.5.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.6.0",
    "mypy>=1.13.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
asurdev = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
